refactor

+++ src/domain/use-case/negocio/delete-negocio.ucase.ts
@@ -1,19 +1,18 @@
-import { INegocioRepository } from '../../repositories';
+import { CustomError } from "../../errors/custom.error";
 
-
-
-
-
-
+import { INegocioRepository } from "../../repositories";
 
 export class DeleteNegocio {
+  constructor(private readonly negocioRepository: INegocioRepository) {}
 
-    constructor(
-        private readonly negocioRepository: INegocioRepository,
-    ){}
+  async execute(id: number): Promise<void> {
+    const existNegocio = this.negocioRepository.verifyNegocioExists(id);
 
-    async execute(id: number): Promise<void> {
-        await this.negocioRepository.deleteNegocio(id);
-    }
+    //todo: validate that user only can delete its own negocio
 
+    if (!existNegocio)
+      throw CustomError.notFound(`Negocio with id: ${id} does not exists.`);
+
+    await this.negocioRepository.deleteNegocio(id);
+  }
 }
